[CHORE] AWS 서버 이관 #424
- selenium test api 모듈 이동
